Reimplement GLibEventLoopPolicy
The 'context' that the policy operates in is the global default
MainContext when in the main thread, and the thread default MainContext
(if it has been set) if in another thread. set_event_loop() is now
understood as setting the "default GLibEventLoop" for the current
context. Multiple GLibEventLoop's can share the same MainContext.